tests: update test data and add test for comment_parser.
